--- demo/pages/IndexList/index.json
@@ -1,8 +1,0 @@
-{
-  "defaultTitle": "Index-List",
-  "pullRefresh": false,
-  "usingComponents": {
-    "index-list": "../../../src/IndexList/index"
-  }
-}
-
