+++ demo/utils/constants.js
@@ -300,11 +300,6 @@ export const componentList = [
         name: 'SafeArea',
         nameZN: '安全区',
         path: '/pages/SafeArea/index',
-      },
-      {
-        name: 'IndexList',
-        nameZN: '索引选择器',
-        path: '/pages/IndexList/index',
       }
     ],
   },
