+++ src/BladeView/index.axml
@@ -1,70 +1,68 @@
+<import-sjs name="util" from="./index.sjs"/>
 
-<view class="amd-blade-view-container">
-  <!-- content -->
-  <scroll-view
-    class="amd-blade-view-body"
-    scroll-y="{{true}}"
-    onScroll="onScroll"
-    scroll-top="{{scrollTop}}"
-  >
-    <view
-      id="amd-blade-view-group-{{group.key}}"
-      class="amd-blade-view-body-group"
-      a:for="{{data}}"
-      a:for-index="groupIndex"
-      a:for-item="group"
-      key="{{group.key}}"
-      data-key="{{group.key}}"
-    >
+<template name="index">
+  <view class="amd-blade-view-sidebar" disable-scroll="{{true}}" hover-stop-propagation>
+    <view a:if="{{data.showMask}}" class="amd-blade-view-sidebar-mask"></view>
+    <view class="amd-blade-view-sidebar">
       <view
-        class="amd-blade-view-body-group-title {{ groupIndex ? '': 'amd-blade-view-body-group-title-first' }} {{sticky && currentKey===group.key ? 'amd-blade-view-body-group-title-sticky' : ''}}"
-        id="amd-blade-view-group-title-{{group.key}}"
+        a:for="{{data.data}}"
+        class="amd-blade-view-sidebar-item {{data.touchKey ? ((data.touchKey === item.key) ? 'amd-blade-view-sidebar-active' : '') : (index === data.currentKey ? 'amd-blade-view-sidebar-active' : '')}}"
+        catchTouchStart="onTouchStart"
+        catchTouchMove="onTouchMove"
+        catchTouchEnd="onTouchEnd"
+        data-item="{{item, index}}"
+        id="{{'i-alphabet-' + index}}"
+        style="{{{ width: `${data.itemHeight}px`, height: `${data.itemHeight}px` }}}"
       >
-        {{group.title}}
-      </view>
-      <view class="amd-blade-view-body-group-item-container">
-        <view
-          class="amd-blade-view-body-group-item"
-          hover-class="amd-blade-view-body-group-item-hover"
-          hover-start-time="20"
-          hover-stay-time="50"
-          a:for="{{group.items}}"
-          data-item="{{item}}"
-          data-group="{{group}}"
-          onTap="onTapItem"
-        >
-          <view class="amd-blade-view-body-group-item-text {{ index ? 'amd-blade-view-body-group-item-text-not-first': '' }} {{'amd-blade-view-body-group-item-text-selected'}}">
-            <slot item="{{item}}">{{item.label}}</slot>
-          </view>
+        {{item.title}}
+        <view class="amd-blade-view-sidebar-item-tip" a:if="{{item.key === data.touchKey}}">
+          {{item.title}}
         </view>
       </view>
     </view>
-  </scroll-view>
-
-  <!-- sidebar -->
-  <view
-    class="amd-blade-view-sidebar"
-    onTouchStart="onTouchStart"
-    onTouchMove="onTouchMove"
-    onTouchEnd="onTouchEnd"
-    disable-scroll="{{true}}"
-  >
-    <view
-      class="amd-blade-view-sidebar-item"
-      a:for="{{data}}"
-      data-key="{{item.key}}"
-      key="{{item.key}}"
-      onTap="onTapSideItem"
-    >
-      <view
-        a:if="{{isTouchingSideBar&&currentKey===item.key}}"
-        class="amd-blade-view-sidebar-current"
-      >
-        {{currentKey}}
-      </view>
-      <view class="amd-blade-view-sidebar-item-text {{currentKey===item.key ? 'active' : ''}}">
-        {{item.key}}
-      </view>
-    </view>
   </view>
+</template>
+
+<view class="amd-blade-view {{className}}" a:if="{{type === 'base' || type === 'diy'}}">
+  <scroll-view
+    scroll-y="true"
+    scroll-into-view="{{scrollId}}"
+    class="amd-blade-view"
+    onScroll="onScroll"
+  >
+    <view a:for="{{data}}" class="amd-blade-view-item" id="{{`amd-blade-view-list_${item.key}`}}">
+      <view
+        a:if="{{type === 'base'}}"
+        style="{{{ position: sticky ? 'sticky' : 'relative' }}}"
+        class="list-index-item-s amd-blade-view-item-title {{amd-index-list-ley}}"
+      >
+        {{ item.title }}
+      </view>
+      <view
+        a:if="{{type === 'base'}}"
+        a:for="{{item.items}}"
+        class="amd-blade-view-item-group {{amd-index-list-item}}"
+        id="{{`amd-blade-view-item_${iarr.value}`}}"
+        data-item="{{iarr, item}}"
+        onTap="onChangeGroupItem"
+        a:for-item="iarr"
+      >
+        <view
+          a:if="{{iarr.avator}}"
+          class="amd-blade-view-item_avator"
+          style="{{{ 'background-image': `url(${iarr.avator})` }}}"/>
+        <view class="amd-blade-view-item_name">{{iarr.label}}</view>
+      </view>
+      <view a:if="{{type === 'diy'}}" class="list-index-item-s" >
+        <slot name="header" value="{{ item }}" />
+      </view>
+      <slot a:if="{{type === 'diy'}}" name="footer" value="{{item.items}}"/>
+    </view>
+  </scroll-view>
+  <view class="amd-blade-view-list">
+    <template is="index" data="{{ data: { data, showMask, touchKey, activeKey, itemHeight, currentKey } }}"/>
+  </view>
+</view>
+<view class="amd-blade-view-list" a:if="{{type === 'only'}}">
+  <template is="index" data="{{ data: { data, showMask, touchKey, activeKey, itemHeight, currentKey } }}"/>
 </view>
