+++ src/BladeView/index.json
@@ -1,4 +1,4 @@
 {
   "component": true,
-  "usingComponents": {}
+  "styleIsolation": "apply-shared"
 }
