+++ src/BladeView/index.less
@@ -1,118 +1,123 @@
-@import (reference) "../_base/index.less";
-@import '../style/mixins/hairline.less';
-
-@BladeViewPrefix: amd-blade-view;
-
-.@{BladeViewPrefix} {
-  &-container {
-    position: relative;
-    background: #fff;
-  }
-
-  &-body {
-    height: 100vh;
-
-    &-group {
-      &-title {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        height: 70 * @rpx;
-        padding: 0 24 * @rpx;
-        color: #999;
-        font-size: 26 * @rpx;
-        background: #f5f5f5;
-
-        &-first {
-          border-top: 1 * @rpx solid #eee;
-        }
-
-        &-sticky {
-          position: sticky;
-          top: 0;
-          z-index: 99;
-        }
-      }
-
-      &-item {
-        padding-left: 24 * @rpx;
-
-        &-container {
-          position: relative;
-          .hairline('top');
-          .hairline('bottom');
-        }
-
-        &-text {
-          padding: 24 * @rpx 24 * @rpx 24 * @rpx 0;
-          overflow: hidden;
-          text-overflow: ellipsis;
-          white-space: nowrap;
-          line-height: 48 * @rpx;
-
-          &-not-first {
-            position: relative;
-            .hairline('top');
-          }
-
-          &-selected {
-            color: #3875f6;
-          }
-        }
-
-        &-hover {
-          background: #eee;
-          opacity: 0.7;
+.amd-blade-view-sidebar {
+    min-height: 48px;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    z-index: 5;
+    &-mask {
+      position: fixed;
+      width: 100vw;
+      height: 100vh;
+      left: 0;
+      top: 0;
+      z-index: 4;
+    }
+    .amd-blade-view-sidebar-active {
+      border-radius: 16rpx;
+      background-color: #1677ff;
+      color: white;
+    }
+    &-item {
+      font-size: 11px;
+      font-weight: bold;
+      color: #999;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      position: relative;
+      &-tip {
+        position: absolute;
+        left: -210rpx;
+        bottom: -36rpx;
+        width: 100rpx;
+        height: 100rpx;
+        border-radius: 100rpx;
+        line-height: 100rpx;
+        text-align: center;
+        background: rgb(204, 204, 204);
+        color: white;
+        font-size: 56rpx;
+        &::after {
+          content: "";
+          position: absolute;
+          top: -30rpx;
+          right: -60rpx;
+          display: block;
+          margin: 40rpx auto;
+          width: 0;
+          height: 0;
+          border: 40rpx solid transparent;
+          border-left-color: rgb(204, 204, 204);
         }
       }
     }
+}
+.amd-blade-view {
+  width: 100%;
+  height: 100%;
+  position: relative;
+  &-item {
+    width: 100%;
+    height: auto;
+    &-title {
+      height: 74rpx;
+      width: calc(100% - 24rpx);
+      background: #f5f5f5;
+      padding-left: 24rpx;
+      line-height: 74rpx;
+      color: #999;
+      position: sticky;
+      top: -1rpx;
+      left: 0;
+      z-index: 2;
+    }
+    &-group {
+      height: 96rpx;
+      width: 100%;
+      background: #fff;
+      display: flex;
+      flex-direction: row;
+      align-items: center;
+      position: relative;
+      z-index: 1;
+      &::after {
+        /* stylelint-disable-next-line string-quotes */
+        content: '';
+        position: absolute;
+        height: 1px;
+        max-height: 1px;
+        min-height: 1px;
+        transform: scaleY(0.5);
+        background-color: #e5e5e5;
+        bottom: 0;
+        width: calc(100% - 24rpx);
+        right: 0;
+      }
+    }
+    &_avator {
+      width: 64rpx;
+      height: 64rpx;
+      background-repeat: no-repeat;
+      background-position: center;
+      background-size: 100%;
+      border-radius: 8rpx;
+      margin-left: 24rpx;
+    }
+    &_name {
+      width: calc(100% - 200rpx);
+      margin-left: 24rpx;
+      height: 100%;
+      display: flex;
+      align-items: center;
+      font-size: 34rpx;
+      color: #333;
+    }
   }
-
-  &-sidebar {
+  &-list {
     position: absolute;
+    right: 16rpx;
     top: 50%;
     transform: translateY(-50%);
-    right: 0;
-    overflow: visible;
-    z-index: 99;
-
-    &-item {
-      width: 96 * @rpx;
-      position: relative;
-      right: 16 * @rpx;
-
-      &-text {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        width: 32 * @rpx;
-        height: 32 * @rpx;
-        border-radius: 32 * @rpx;
-        color: #999;
-        font-size: 22 * @rpx;
-        margin-left: 64 * @rpx;
-
-        &.active {
-          color: #fff;
-          border-radius: 32 * @rpx;
-          background: #3875f6;
-        }
-      }
-    }
-
-    &-current {
-      position: absolute;
-      top: 50%;
-      transform: translateY(-50%);
-      right: 96 * @rpx;
-      width: 96 * @rpx;
-      height: 96 * @rpx;
-      color: #fff;
-      font-size: 50 * @rpx;
-      line-height: 96 * @rpx;
-      text-align: center;
-      background: #ccc;
-      border-radius: 50%;
-    }
+    z-index: 3;
   }
 }
