+++ src/BladeView/props.js
@@ -1,4 +1,11 @@
 export const BladeViewDefaultProps = {
+  type: 'base',
   data: [],
-  sticky: false,
-};
+  sticky: true,
+  value: '',
+  scrollToKey: '',
+  vibrate: true,
+  className: '',
+  onChange: () => {},
+  onChangeSwipeKey: () => {},
+}
