--- src/IndexList/index.json
@@ -1,4 +1,0 @@
-{
-  "component": true,
-  "styleIsolation": "apply-shared"
-}
