--- src/IndexList/index.less
@@ -1,123 +1,0 @@
-.i-alphabet {
-    min-height: 48px;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    z-index: 5;
-    &-mask {
-      position: fixed;
-      width: 100vw;
-      height: 100vh;
-      left: 0;
-      top: 0;
-      z-index: 4;
-    }
-    .active-key {
-      border-radius: 16rpx;
-      background-color: #1677ff;
-      color: white;
-    }
-    &-item {
-      font-size: 11px;
-      font-weight: bold;
-      color: #999;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      position: relative;
-      &-tip {
-        position: absolute;
-        left: -210rpx;
-        bottom: -36rpx;
-        width: 100rpx;
-        height: 100rpx;
-        border-radius: 100rpx;
-        line-height: 100rpx;
-        text-align: center;
-        background: rgb(204, 204, 204);
-        color: white;
-        font-size: 56rpx;
-        &::after {
-          content: "";
-          position: absolute;
-          top: -30rpx;
-          right: -60rpx;
-          display: block;
-          margin: 40rpx auto;
-          width: 0;
-          height: 0;
-          border: 40rpx solid transparent;
-          border-left-color: rgb(204, 204, 204);
-        }
-      }
-    }
-}
-.base-index {
-  width: 100%;
-  height: 100%;
-  position: relative;
-  &-item {
-    width: 100%;
-    height: auto;
-    &_index {
-      height: 74rpx;
-      width: calc(100% - 24rpx);
-      background: #f5f5f5;
-      padding-left: 24rpx;
-      line-height: 74rpx;
-      color: #999;
-      position: sticky;
-      top: -1rpx;
-      left: 0;
-      z-index: 2;
-    }
-    &_con {
-      height: 96rpx;
-      width: 100%;
-      background: #fff;
-      display: flex;
-      flex-direction: row;
-      align-items: center;
-      position: relative;
-      z-index: 1;
-      &::after {
-        /* stylelint-disable-next-line string-quotes */
-        content: '';
-        position: absolute;
-        height: 1px;
-        max-height: 1px;
-        min-height: 1px;
-        transform: scaleY(0.5);
-        background-color: #e5e5e5;
-        bottom: 0;
-        width: calc(100% - 24rpx);
-        right: 0;
-      }
-    }
-    &_avator {
-      width: 64rpx;
-      height: 64rpx;
-      background-repeat: no-repeat;
-      background-position: center;
-      background-size: 100%;
-      border-radius: 8rpx;
-      margin-left: 24rpx;
-    }
-    &_name {
-      width: calc(100% - 200rpx);
-      margin-left: 24rpx;
-      height: 100%;
-      display: flex;
-      align-items: center;
-      font-size: 34rpx;
-      color: #333;
-    }
-  }
-  &-list {
-    position: absolute;
-    right: 16rpx;
-    top: 50%;
-    transform: translateY(-50%);
-    z-index: 3;
-  }
-}
