--- src/IndexList/props.js
@@ -1,12 +1,0 @@
-export const IndexListDefaultProps = {
-  type: 'base',
-  indexList: [],
-  indexArr: [],
-  sticky: true,
-  vibrate: true,
-  'amd-index-list-index': '',
-  'amd-index-list-item': '',
-  'amd-index-list-list': '',
-  'amd-index-list-key': '',
-  onChange: () => {},
-}
